split getconfV2.go in files, renamed and purged old files
